pyproject.toml
[tool.poetry]
name = "tile-renderer"
version = "4.0.0"
description = "Leaflet.js streetmap tile renderer"
authors = ["7d <i.third.7d@protonmail.com>"]
license = "gpl-3.0"
readme = "README.md"
repository = "https://github.com/MRT-Map/tile-renderer"
documentation = "https://tile-renderer.rtfd.io"
keywords = ['leaflet', 'leaflet.js', 'leafletjs', 'map', 'tiles', 'renderer', 'tile-renderer', 'mapping']
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
packages = [{ include = "renderer" }]


[tool.poetry.dependencies]
python = "^3.9,<3.12"
Pillow = "^9.2.0"
schema = "^0.7.5"
vector = "^1.0.0"
psutil = "^5.9.5"
ImageHash = "^4.2.1"
fonttools = "^4.34.4"
rich = "^13.3.1"
ray = "^2.3.0"
toml = "^0.10.2"
Shapely = "^2.0a1"
msgspec = "^0.14.1"
dill = "^0.3.5.1"
six = "^1.16.0"
methodtools = "^0.4.7"
typing_extensions = "^4.5.0"

[tool.poetry.group.dev.dependencies]
wheel = "^0.37.1"
pip = "^22.1.2"
black = "^23.1.0"
pytest = "^7.1.3"
mypy = "^1.0.0"
isort = "^5.12.0"
setuptools = "^65.6.3"
types-toml = "^0.10.8.1"
types-psutil = "^5.9.5.6"
types-pillow = "^9.4.0.5"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"
sphinxcontrib-programoutput = "^0.17"
sphinx-codeautolink = "^0.14.1"
ruff = "^0.0.261"

[tool.poetry.scripts]
renderer = 'renderer.__main__:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"

[[tool.mypy.overrides]]
module = "ray.*"
ignore_errors = true

[tool.ruff]
select = ["E", "F", "W", "N", "U", "YTT", "ANN", "S", "BLE", "B", "A", "COM", "C4", "DTZ", "EM", "EXE",
    "ISC", "ICN", "G", "INP", "PIE", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TCH", "INT", "ARG", "PTH", "ERA",
    "PGH", "PL", "TRY", "RUF"]
ignore = ["E501", "ANN401", "A003", "ANN101", "ANN102", "SLF001", "PLR2004", "PLR0913", "TRY003", "EM101", "EM102"]